# time: implement the encoding.(Binary|Text)Appender for Time