--- a/doc/next/6-stdlib/99-minor/time/62384.md
+++ b/doc/next/6-stdlib/99-minor/time/62384.md
@@ -1,0 +1,1 @@
+[Time] now implements the [encoding.BinaryAppender] and [encoding.TextAppender] interfaces.